requirements.txt
# Core dependencies
Pillow>=10.0.0
python-barcode>=0.16.0

# API clients
requests>=2.31.0
openai>=1.68.0
pydantic>=2.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Thermal Printer Support
python-escpos>=3.0.0
pyusb>=1.0.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3

# Web Console
Flask>=3.0.0
